feat: options handling, cleanup for plguin register
